Fix multiple-definition and type-mismatch linker errors
- src/hazard/hazrd4.h: remove 'nrow' from HAZRD4 definitions block;
  nrow is already defined in hazard.c via hazard.h (EXTERN int prnt,nrow)
  so having it here too caused a multiple definition error on strict linkers

- src/hazard/hazard_l.c: change 'int yylnctr,yysynerr' to
  'extern int yylnctr; int yysynerr;' — yylnctr is already defined in
  hazard.c via structures.h (EXTERN int yylnctr with HAZARD defined);
  yysynerr remains a definition here as it is unique to the lexer

- src/hazard/setstat.c + setstat.h: unify function signature to
  int* for all three pointer params; the implementation had
  (long*,long*,int*) while the header had (int*,int*,long*) — both wrong.
  All callers pass int* (H->phase1 is int, iearly/iconst/ilate are int*,
  iexc/icxc/ilxc are int), so int* throughout is correct.
  The mismatched signature was an error on Windows/MinGW, warning on Linux.

+++ src/hazard/setstat.h
@@ -1,4 +1,4 @@
 #ifndef SETSTAT_H
 #define SETSTAT_H
-void setstat(int i,int idx,int *phasevar,int *itabl,long *ipxc);
+void setstat(int i,int idx,int *phasevar,int *itabl,int *ipxc);
 #endif /* SETSTAT_H */
